rename module _segmentation_mask.py -> _mask.py
