+++ torchvision/prototype/features/__init__.py
@@ -3,4 +3,4 @@ from ._encoded import EncodedData, EncodedImage, EncodedVideo
 from ._feature import _Feature, is_simple_tensor
 from ._image import ColorSpace, Image
 from ._label import Label, OneHotLabel
-from ._segmentation_mask import Mask
+from ._mask import Mask
